requirements.txt
anyio
fastapi==0.104.1
flask
rembg
pillow
gunicorn
